optimize via new shared/ui SummaryItem component
