+++ src/features/ChallengeStatistics/ChallengeStatistics.module.scss
