+++ src/features/index.ts
@@ -7,3 +7,4 @@ export {
   MakeActiveChallengeButton,
 } from './ManageChallenge/ManageChallenge';
 export { BarChart, BarStackedChart, BarStackedSingleChart } from './Charts';
+export { ChallengeStatistics } from './ChallengeStatistics/ChallengeStatistics';
